requirements.txt
setuptools>=65.0
openai-whisper
